reduce repeated stuff, clean
